requirements.txt
requests>=2.31.0
numpy>=1.24.3
json
tabulate>=0.9.0
python-dotenv>=1.0.0
uuid>=1.30
hashlib3>=20081119